bugfix: corrected annotations and tests
