+++ src/main/java/ru/yandex/practicum/javafilmorate/JavaFilmorateApplication.java
@@ -1,7 +1,5 @@
 package ru.yandex.practicum.javafilmorate;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
