+++ src/main/java/ru/yandex/practicum/javafilmorate/service/UserService.java
@@ -2,11 +2,14 @@ package ru.yandex.practicum.javafilmorate.service;
 
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.javafilmorate.exceptions.EntityAlreadyExistsException;
 import ru.yandex.practicum.javafilmorate.exceptions.EntityDoesNotExistException;
+import ru.yandex.practicum.javafilmorate.exceptions.InvalidPathVariableOrParameterException;
 import ru.yandex.practicum.javafilmorate.model.User;
 import ru.yandex.practicum.javafilmorate.storage.UserStorage;
+import ru.yandex.practicum.javafilmorate.util.StringValidator;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -16,11 +19,13 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class UserService {
 
     private final UserStorage userStorage;
 
     public User addNewUserToStorage(User user) {
+        setUserNameIfNeeded(user);
         return userStorage.addUser(user);
     }
 
@@ -37,10 +42,13 @@ public class UserService {
     }
 
     public User getUserFromStorage(long id) {
+        checkUserExistence(id);
         return userStorage.getUser(id);
     }
 
     public void addUsersToFriendsInStorage(long id, long friendId) {
+        checkUserExistence(id);
+        checkUserExistence(friendId);
         User user = userStorage.getUser(id);
         if (user.getFriends() == null) {
             user.setFriends(new HashSet<>());
@@ -60,6 +68,8 @@ public class UserService {
     }
 
     public void deleteUsersFromFriendsInStorage(long id, long friendId) {
+        checkUserExistence(id);
+        checkUserExistence(friendId);
         User user = userStorage.getUser(id);
         if (user.getFriends() == null) {
             user.setFriends(new HashSet<>());
@@ -79,7 +89,7 @@ public class UserService {
     }
 
     public List<User> getUserFriendsFromStorage(long id) {
-
+        checkUserExistence(id);
         Set<Long> userFriends = userStorage.getUser(id).getFriends();
 
         if (userFriends == null) {
@@ -92,7 +102,8 @@ public class UserService {
     }
 
     public List<User> getUsersCommonFriendsFromStorage(long id, long otherId) {
-
+        checkUserExistence(id);
+        checkUserExistence(otherId);
         Set<Long> userFriends = userStorage.getUser(id).getFriends();
 
         if (userFriends == null) {
@@ -106,5 +117,29 @@ public class UserService {
                 .filter(friendsFriends::contains)
                 .map(userStorage::getUser)
                 .collect(Collectors.toList());
+    }
+
+    private void checkUserExistence(long id) {
+
+        if (id < 1) {
+            String exceptionMessage = String.format("Пользователь с  id = %d не может существовать", id);
+            log.warn("Ошибка при запросе пользователя. Сообщение исключения: {}",
+                    exceptionMessage);
+            throw new InvalidPathVariableOrParameterException("id", exceptionMessage);
+        }
+
+        if (!doesUserExistInStorage(id)) {
+            String exceptionMessage = String.format("Пользователь с  id = %d не существует", id);
+            log.warn("Ошибка при запросе пользователя. Сообщение исключения: {}",
+                    exceptionMessage);
+            throw new EntityDoesNotExistException(exceptionMessage);
+        }
+    }
+
+    private void setUserNameIfNeeded(User user) {
+
+        if (StringValidator.isNullOrEmpty(user.getName())) {
+            user.setName(user.getLogin());
+        }
     }
 }
