+++ src/test/java/ru/yandex/practicum/javafilmorate/controllers/FilmControllerTest.java
@@ -2,13 +2,8 @@ package ru.yandex.practicum.javafilmorate.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.SpringApplication;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.ConfigurableApplicationContext;
-import ru.yandex.practicum.javafilmorate.JavaFilmorateApplication;
 import ru.yandex.practicum.javafilmorate.model.Film;
 
 import java.io.IOException;
@@ -18,71 +13,29 @@ import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.time.Duration;
 import java.time.LocalDate;
+import java.util.Random;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-@SpringBootTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 class FilmControllerTest {
 
-    private static ConfigurableApplicationContext app;
     private static final int PORT = 8080;
 
     private final ObjectMapper objectMapper = new ObjectMapper()
             .findAndRegisterModules()
             .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
-    private static Film titanicWithEmptyName;
-    private static Film titanicWithTooLongDescription;
-    private static Film titanicWithNegativeDuration;
-    private static Film titanicWithTooOldReleaseDate;
+    @Test
+    public void shouldReturnFilmNameNullOrEmptyException() throws IOException, InterruptedException {
 
-    @BeforeAll
-    static void beforeAll() {
-
-        app = SpringApplication.run(JavaFilmorateApplication.class);
-
-        titanicWithEmptyName = Film.builder().id(1)
+        final Film titanicWithEmptyName = Film.builder().id(1)
                 .name("")
                 .description("Test description")
                 .duration(Duration.ofMinutes(90))
                 .releaseDate(LocalDate.of(1997, 1, 23))
                 .build();
 
-        String tooLongDescription = "Test description Test description Test description" +
-                " Test description Test description Test description" +
-                "Test description Test description Test description" +
-                "Test description Test description Test description" +
-                "Test description Test description Test description";
-
-        titanicWithTooLongDescription = Film.builder().id(1)
-                .name("Titanic")
-                .description(tooLongDescription)
-                .duration(Duration.ofMinutes(90))
-                .releaseDate(LocalDate.of(1997, 1, 23))
-                .build();
-
-        titanicWithNegativeDuration = Film.builder().id(1)
-                .name("Titanic")
-                .description("Test Description")
-                .duration(Duration.ofMinutes(-90))
-                .releaseDate(LocalDate.of(1997, 1, 23))
-                .build();
-
-        titanicWithTooOldReleaseDate = Film.builder().id(1)
-                .name("Titanic")
-                .description("Test Description")
-                .duration(Duration.ofMinutes(90))
-                .releaseDate(LocalDate.of(1797, 1, 23))
-                .build();
-    }
-
-    @AfterAll
-    static void afterAll() {
-        app.close();
-    }
-
-    @Test
-    public void shouldReturnFilmNameNullOrEmptyException() throws IOException, InterruptedException {
         HttpClient client = HttpClient.newHttpClient();
         URI uri = URI.create("http://localhost:" + PORT + "/films");
         String requestBody = objectMapper.writeValueAsString(titanicWithEmptyName);
@@ -100,6 +53,14 @@ class FilmControllerTest {
 
     @Test
     public void shouldReturnFilmDescriptionTooLongException() throws IOException, InterruptedException {
+
+        final Film titanicWithTooLongDescription = Film.builder().id(1)
+                .name("Titanic")
+                .description(generateLetterString(201))
+                .duration(Duration.ofMinutes(90))
+                .releaseDate(LocalDate.of(1997, 1, 23))
+                .build();
+
         HttpClient client = HttpClient.newHttpClient();
         URI uri = URI.create("http://localhost:" + PORT + "/films");
         String requestBody = objectMapper.writeValueAsString(titanicWithTooLongDescription);
@@ -117,6 +78,14 @@ class FilmControllerTest {
 
     @Test
     public void shouldReturnFilmDurationIsNegativeValueException() throws IOException, InterruptedException {
+
+        final Film titanicWithNegativeDuration = Film.builder().id(1)
+                .name("Titanic")
+                .description("Test Description")
+                .duration(Duration.ofMinutes(-90))
+                .releaseDate(LocalDate.of(1997, 1, 23))
+                .build();
+
         HttpClient client = HttpClient.newHttpClient();
         URI uri = URI.create("http://localhost:" + PORT + "/films");
         String requestBody = objectMapper.writeValueAsString(titanicWithNegativeDuration);
@@ -134,6 +103,14 @@ class FilmControllerTest {
 
     @Test
     public void shouldReturnFilmReleaseDateException() throws IOException, InterruptedException {
+
+        final Film titanicWithTooOldReleaseDate = Film.builder().id(1)
+                .name("Titanic")
+                .description("Test Description")
+                .duration(Duration.ofMinutes(90))
+                .releaseDate(LocalDate.of(1797, 1, 23))
+                .build();
+
         HttpClient client = HttpClient.newHttpClient();
         URI uri = URI.create("http://localhost:" + PORT + "/films");
         String requestBody = objectMapper.writeValueAsString(titanicWithTooOldReleaseDate);
@@ -149,5 +126,15 @@ class FilmControllerTest {
         assertEquals(400, response.statusCode());
     }
 
+    String generateLetterString(int targetStringLength) {
+        int leftLimit = 97;
+        int rightLimit = 122;
+        Random random = new Random();
+
+        return random.ints(leftLimit, rightLimit + 1)
+                .limit(targetStringLength)
+                .collect(StringBuilder::new, StringBuilder::appendCodePoint, StringBuilder::append)
+                .toString();
+    }
 
 }
