+++ src/test/java/ru/yandex/practicum/javafilmorate/controllers/UserControllerTest.java
@@ -2,13 +2,9 @@ package ru.yandex.practicum.javafilmorate.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.SpringApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.ConfigurableApplicationContext;
-import ru.yandex.practicum.javafilmorate.JavaFilmorateApplication;
 import ru.yandex.practicum.javafilmorate.model.User;
 
 import java.io.IOException;
@@ -20,7 +16,7 @@ import java.time.LocalDate;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-@SpringBootTest
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 class UserControllerTest {
 
     private static ConfigurableApplicationContext app;
@@ -30,69 +26,16 @@ class UserControllerTest {
 
     private static final int PORT = 8080;
 
-    private static User userWithoutEmail;
+    @Test
+    public void shouldReturnUserEmailNullOrEmptyException() throws IOException, InterruptedException {
 
-    private static User userWithIncorrectEmail;
-
-    private static User userWithEmptyLogin;
-
-    private static User userWithBlankInLogin;
-
-    private static User userWithBirthdayInTheFuture;
-
-    @BeforeAll
-    static void beforeAll() {
-
-        app = SpringApplication.run(JavaFilmorateApplication.class);
-
-        userWithoutEmail = User.builder()
+        final User userWithoutEmail = User.builder()
                 .id(1)
                 .email(null)
                 .login("TestLogin")
                 .name("TestName")
                 .birthday(LocalDate.of(1993, 2, 28))
                 .build();
-
-        userWithIncorrectEmail = User.builder()
-                .id(1)
-                .email("Test.mail")
-                .login("TestLogin")
-                .name("TestName")
-                .birthday(LocalDate.of(1993, 2, 28))
-                .build();
-
-        userWithEmptyLogin = User.builder()
-                .id(1)
-                .email("Test@mail.com")
-                .login("")
-                .name("TestName")
-                .birthday(LocalDate.of(1993, 2, 28))
-                .build();
-
-        userWithBlankInLogin = User.builder()
-                .id(1)
-                .email("Test@mail.com")
-                .login("Test Login")
-                .name("TestName")
-                .birthday(LocalDate.of(1993, 2, 28))
-                .build();
-
-        userWithBirthdayInTheFuture = User.builder()
-                .id(1)
-                .email("Test@mail.com")
-                .login("TestLogin")
-                .name("TestName")
-                .birthday(LocalDate.of(2100, 2, 28))
-                .build();
-    }
-
-    @AfterAll
-    static void afterAll() {
-        app.close();
-    }
-
-    @Test
-    public void shouldReturnUserEmailNullOrEmptyException() throws IOException, InterruptedException {
 
         HttpClient client = HttpClient.newHttpClient();
         URI uri = URI.create("http://localhost:" + PORT + "/users");
@@ -111,6 +54,14 @@ class UserControllerTest {
 
     @Test
     public void shouldReturnUserEmailIsIncorrectException() throws IOException, InterruptedException {
+
+        final User userWithIncorrectEmail = User.builder()
+                .id(1)
+                .email("Test.mail")
+                .login("TestLogin")
+                .name("TestName")
+                .birthday(LocalDate.of(1993, 2, 28))
+                .build();
 
 
         HttpClient client = HttpClient.newHttpClient();
@@ -131,6 +82,15 @@ class UserControllerTest {
     @Test
     public void shouldReturnUserLoginIsEmptyOrBlankException() throws IOException, InterruptedException {
 
+
+        final User userWithEmptyLogin = User.builder()
+                .id(1)
+                .email("Test@mail.com")
+                .login("")
+                .name("TestName")
+                .birthday(LocalDate.of(1993, 2, 28))
+                .build();
+
         HttpClient client = HttpClient.newHttpClient();
         URI uri = URI.create("http://localhost:" + PORT + "/users");
         String requestBody = objectMapper.writeValueAsString(userWithEmptyLogin);
@@ -144,11 +104,19 @@ class UserControllerTest {
                 .build();
         HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
         assertEquals(400, response.statusCode());
-
     }
 
     @Test
     public void shouldReturnUserLoginIsEmptyOrBlankExceptionBecauseOfBlank() throws IOException, InterruptedException {
+
+        final User userWithBlankInLogin = User.builder()
+                .id(1)
+                .email("Test@mail.com")
+                .login("Test Login")
+                .name("TestName")
+                .birthday(LocalDate.of(1993, 2, 28))
+                .build();
+
 
         HttpClient client = HttpClient.newHttpClient();
         URI uri = URI.create("http://localhost:" + PORT + "/users");
@@ -168,6 +136,14 @@ class UserControllerTest {
     @Test
     public void shouldReturnUserBirthdayException() throws IOException, InterruptedException {
 
+        final User userWithBirthdayInTheFuture = User.builder()
+                .id(1)
+                .email("Test@mail.com")
+                .login("TestLogin")
+                .name("TestName")
+                .birthday(LocalDate.of(2100, 2, 28))
+                .build();
+
         HttpClient client = HttpClient.newHttpClient();
         URI uri = URI.create("http://localhost:" + PORT + "/users");
         String requestBody = objectMapper.writeValueAsString(userWithBirthdayInTheFuture);
@@ -182,5 +158,4 @@ class UserControllerTest {
         HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
         assertEquals(400, response.statusCode());
     }
-
 }
